Kinova API extracted into singleton
