+++ HumanUnityProjectV0_3/Assets/KinovaAPI.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 21078d91013edcf4a992504e098cc71f
+timeCreated: 1501144872
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
